feat: add build scripts / reorganize src directory
