+++ package.json
@@ -3,8 +3,10 @@
   "private": true,
   "version": "0.0.0",
   "scripts": {
-    "develop": "npm run storybook",
-    "build": "tsc && vite build",
+    "develop": "run-p watch-typescript storybook",
+    "clean": "rm -rf dist",
+    "build": "npm run clean && find src/components -name '*.ts' -not -name '*.stories.ts' -exec npx vite build -- {} \\;",
+    "watch-typescript": "tsc --watch",
     "storybook": "storybook dev -p 6006",
     "storybook:build": "storybook build",
     "storybook:deploy": "storybook-to-ghpages -o .out",
@@ -27,6 +29,7 @@
     "@storybook/blocks": "^7.4.6",
     "@storybook/web-components": "^7.4.6",
     "@storybook/web-components-vite": "^7.4.6",
+    "npm-run-all": "^4.1.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "storybook": "^7.4.6",
