+++ src/tokens/components/border-tokens/border-tokens.stories.ts
@@ -1,24 +1,24 @@
 import type { Meta, StoryObj } from '@storybook/web-components';
-import tokens from '../../../tokens/transformed.tokens.json';
 import figma from '../../../../.storybook/configma.json';
+import tokens from '../../transformed.tokens.json';
 
-import './size-tokens';
+import './border-tokens';
 
 const meta: Meta = {
-  title: 'Tokens/Size',
-  component: 'size-tokens',
+  title: 'Tokens/Border',
+  component: 'border-tokens',
 };
 
 export default meta;
 
-export const Size: StoryObj = {
+export const Border: StoryObj = {
   args: {
-    tokens: tokens.size,
+    tokens: tokens.border,
   },
   parameters: {
     design: {
       type: 'figma',
-      url: `${figma.url}${figma.size}`,
+      url: `${figma.url}${figma.border}`,
     },
   },
 };
