+++ src/tokens/components/size-tokens/size-tokens.stories.ts
@@ -1,24 +1,24 @@
 import type { Meta, StoryObj } from '@storybook/web-components';
-import tokens from '../../../tokens/transformed.tokens.json';
 import figma from '../../../../.storybook/configma.json';
+import tokens from '../../transformed.tokens.json';
 
-import './border-tokens';
+import './size-tokens';
 
 const meta: Meta = {
-  title: 'Tokens/Border',
-  component: 'border-tokens',
+  title: 'Tokens/Size',
+  component: 'size-tokens',
 };
 
 export default meta;
 
-export const Border: StoryObj = {
+export const Size: StoryObj = {
   args: {
-    tokens: tokens.border,
+    tokens: tokens.size,
   },
   parameters: {
     design: {
       type: 'figma',
-      url: `${figma.url}${figma.border}`,
+      url: `${figma.url}${figma.size}`,
     },
   },
 };
